Move var handle adapters from MethodHandles into incubator module (MemoryHandles)
